refactor: move VerifyChainLock and ChainLockSig between modules to untangle chain-state dependencies
chain-state should be available to validate chainlocks without using
chainlock/handlers.cpp which is tightly connected to network code.

On other hand, VerifyChainLock uses quorumsman, so it can't be just moved to chainlock/clsig or something
like that because it will create new circular dependency over quorumsman.

Though, chainlock::Chainlock and chainlock::ChainLockSig are tightly connected
and there is no usages of ChainLockSig withdout Chainlock except regression tests;
moving ChainLockSig to chainlock.h is fine then.
