+++ src/chainlock/chainlock.h
@@ -5,18 +5,44 @@
 #ifndef BITCOIN_CHAINLOCK_CHAINLOCK_H
 #define BITCOIN_CHAINLOCK_CHAINLOCK_H
 
-#include <chain.h>
-#include <chainlock/clsig.h>
+#include <bls/bls.h>
 #include <gsl/pointers.h>
+#include <serialize.h>
 #include <sync.h>
 #include <uint256.h>
 
+class CBlockIndex;
 class CSporkManager;
+class uint256;
 
 namespace chainlock {
 
 //! Depth of block including transactions before it's considered safe
 static constexpr int32_t TX_CONFIRM_THRESHOLD{5};
+
+struct ChainLockSig {
+private:
+    int32_t nHeight{-1};
+    uint256 blockHash;
+    CBLSSignature sig;
+
+public:
+    ChainLockSig();
+    ~ChainLockSig();
+
+    ChainLockSig(int32_t nHeight, const uint256& blockHash, const CBLSSignature& sig);
+
+    [[nodiscard]] int32_t getHeight() const { return nHeight; }
+    [[nodiscard]] const uint256& getBlockHash() const { return blockHash; }
+    [[nodiscard]] const CBLSSignature& getSig() const { return sig; }
+    [[nodiscard]] bool IsNull() const { return nHeight == -1 && blockHash == uint256(); }
+    [[nodiscard]] std::string ToString() const;
+
+    SERIALIZE_METHODS(ChainLockSig, obj)
+    {
+        READWRITE(obj.nHeight, obj.blockHash, obj.sig);
+    }
+};
 
 class Chainlocks
 {
